Fix registry auth to not be docker.io specific.
To authenticate with Gitlab, we need to:

- Use a custom service name (does not match the Domain in Gitlab).
- Not presume the path of the auth url.

This is backwards-incompatible, because the required url for auth_service_url
changes (needs to include the path), and because auth_service_name is now required, too.

Finally, the submit an API reques with the received token, we must not attach
basic authentication, otherwise the registry will not accept the bearer token given via Header.